requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0
asyncpg
aiosqlite
psycopg2-binary
python-dotenv
pydantic>=2.0
alembic
